requirements.txt
pytest==7.3.1
selenium==4.9.1
allure-pytest==2.13.2
pytest-xdist==3.3.1
requests==2.31.0